% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSM_9_35.R
\name{HSM_9_35}
\alias{HSM_9_35}
\title{2010 HSM Section 9-35 Site Specific Empirical Bayes Analysis}
\usage{
HSM_9_35(data, group, segment)
}
\arguments{
\item{data}{Input data set. Requires user to specify the safety performance function, observed crashes, and AADT. Safety Performance Function title names are specified according to the titles provided in the FHWA "The Calibrator". SPF_table.RData provides available spf functions. However, segment SPFs other than RUR2U-KABCO have not been coded with version 1.0. Please Note column titles are not dynamic. Use column headings provided in the X9_10_template.}

\item{group}{Input parameter for Step 14. No default is specified. User specifies TRUE or FALSE for Step 14, respectively.}

\item{segment}{Input parameter for segments. No default is specified. User specifies TRUE for Segments or FALSE. Note if segments are used the length (L) column must be populated (units are in miles).}
}
\value{
A tibble
}
\description{
2010 HSM Section 9-35 Site Specific Empirical Bayes Analysis
}
\examples{
HSM_9_35(data=X9_10_seg,group=FALSE,segment = TRUE) # Only the HSM example 9.10 SPF HSM-RUR2U-KABCO is available with version 1.0.

HSM_9_35(data=X9_10_seg,group=TRUE,segment = TRUE)

HSM_9_35(data=X9_10_int,group=FALSE,segment = FALSE)# All FI and Total crash intersection SPF are available with version 1.0.

HSM_9_35(data=X9_10_int,group=TRUE,segment = FALSE)

}
