% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{path.find}
\alias{path.find}
\title{Find all path graphs originated from a given root.}
\usage{
path.find(index, map)
}
\arguments{
\item{index}{Index of a root node (a node whose index never appears in
\code{map[, 2]}).}

\item{map}{Matrix of \code{n.edges}-by-\code{2} dimension, where \code{n.edges}
is the number of directed edges in DAG. The first column has indices of
nodes that edges directing from, whereas the second column gives the indices
of nodes the corresponding edges directing towards.}
}
\value{
Returns a list of path graphs originated from root \code{index}, for
which the \code{i}th element of the returned list is a vector of indices of
nodes in the \code{i}th path graph.
}
\description{
Recursively find all possible path graphs originated from a given root in
DAG.
}
